requirements.txt
weaviate-client>=4.0.0
openai
pydantic
python-multipart
unstructured
python-docx
PyMuPDF
fastapi==0.100.0
pydantic==2.4.0
python-dotenv==1.0.0
motor==3.2.0
pymongo==4.5.0
langchain==0.0.226
langchain-community==0.0.5
langchain-huggingface==0.0.2
bson==0.5.10
bcrypt==4.0.1
docx==0.2.4
openai==0.29.0
uvicorn==0.23.0  # For running FastAPI applications